perf: move theme parsing out of natives to prevent duplicate dependencies
