+++ src/VencordNative.ts
@@ -5,7 +5,6 @@
  */
 
 import { IpcEvents } from "@utils/IpcEvents";
-import type { ThemeHeader } from "@utils/themes";
 import { IpcRes } from "@utils/types";
 import { ipcRenderer } from "electron";
 
@@ -22,7 +21,7 @@ export default {
         uploadTheme: (fileName: string, fileData: string) => invoke<void>(IpcEvents.UPLOAD_THEME, fileName, fileData),
         deleteTheme: (fileName: string) => invoke<void>(IpcEvents.DELETE_THEME, fileName),
         getThemesDir: () => invoke<string>(IpcEvents.GET_THEMES_DIR),
-        getThemesList: () => invoke<Array<ThemeHeader>>(IpcEvents.GET_THEMES_LIST),
+        getThemesList: () => invoke<Array<{ fileName: string; content: string; }>>(IpcEvents.GET_THEMES_LIST),
         getThemeData: (fileName: string) => invoke<string | undefined>(IpcEvents.GET_THEME_DATA, fileName)
     },
