+++ src/utils/themes/usercss/index.ts
@@ -9,7 +9,7 @@ import { parse as originalParse, UserstyleHeader } from "usercss-meta";
 
 const UserCSSLogger = new Logger("UserCSS", "#d2acf5");
 
-export function parse(text: string, fileName: string): UserstyleHeader {
+export function usercssParse(text: string, fileName: string): UserstyleHeader {
     var { metadata, errors } = originalParse(text.replace(/\r/g, ""), { allowErrors: true });
 
     if (errors.length) {
